Add an API to deal with the ping file descriptor
The code gluing the protocol to the uloop is mostly open coded in
`ping.c`. To prepare moving this code out of `ping.c` add an API
for protocols to manager their file descriptor without directly
dealing with the private `ufd` member.

While at it also fix the condition used to check for a valid file
descriptor. The current code assume that zero is not a valid file
descriptor which is not correct. As we now always register our
fd right away in uloop we can instead check if it is registered.