Add an ARP pinger
This new protocol use ARP probes to check a target. It is mostly
useful when used on an interface that doesn't have a configured address.
For example on an access point that itself doesn't have an address
on the bridge between the wired and wireless network. Using ARP ping
it is then possible to check if the gateway is reachable.